Remove cell code
